Dispose native integration on application exit
Summary:
This diff disposes OOM support when a user uses Application.Exit (or overall is about to exit Unity game on iOS".
In addition to that, right now we're able to overall dispose our native integration (ignore crashes in the future)

Test Plan: Build and run it against the latest version of Unity test game.

Reviewers: salex, vlussenburg, rqurashi

Reviewed By: salex, vlussenburg, rqurashi

Differential Revision: https://hack.in.backtrace.io/D4338